+++ CHANGELOG.md
@@ -9,6 +9,7 @@ adheres to [Semantic Versioning][semver].
 - `MonzoAPI()._refresh_oath_token()` now doesn't return anything, replaces
   current token and raises `UnableToRefreshTokenException` when token couldn't
   be refreshed.
+- Client secret is now saved in token file JSON file.
 
 ## [v0.10.1][v0.10.1] - 2017-09-24
 ### Fixed